+++ src/protocols/farmer_protocol.py
@@ -5,7 +5,7 @@ from src.types.hashable.BLSSignature import BLSSignature
 from src.types.proof_of_space import ProofOfSpace
 from src.types.sized_bytes import bytes32
 from src.util.cbor_message import cbor_message
-from src.util.ints import uint32, uint64
+from src.util.ints import uint32, uint64, uint128
 
 
 """
@@ -18,7 +18,7 @@ Protocol between farmer and full node.
 class ProofOfSpaceFinalized:
     challenge_hash: bytes32
     height: uint32
-    weight: uint64
+    weight: uint128
     difficulty: uint64
 
 
@@ -26,7 +26,7 @@ class ProofOfSpaceFinalized:
 @cbor_message
 class ProofOfSpaceArrived:
     previous_challenge_hash: bytes32
-    weight: uint64
+    weight: uint128
     quality_string: bytes32
